Move shared functionality to internal/safelinks
Prepping for further expansion of project.
