+++ cmd/usl/main.go
@@ -12,6 +12,8 @@ package main
 import (
 	"fmt"
 	"os"
+
+	"github.com/atc0005/safelinks/internal/safelinks"
 )
 
 func main() {
@@ -33,7 +35,7 @@ func main() {
 			os.Exit(1)
 		}
 
-		input, err := readURLsFromFile(f)
+		input, err := safelinks.ReadURLsFromFile(f)
 		if err != nil {
 			fmt.Printf("Failed to read URLs from %q: %v\n", cfg.Filename, err)
 			os.Exit(1)
@@ -42,7 +44,7 @@ func main() {
 		inputURLs = input
 
 	default:
-		input, err := processInputAsURL(cfg.URL)
+		input, err := safelinks.ProcessInputAsURL(cfg.URL)
 		if err != nil {
 			fmt.Printf("Failed to parse input as URL: %v\n", err)
 			os.Exit(1)
@@ -51,7 +53,7 @@ func main() {
 		inputURLs = input
 	}
 
-	hasErr := processInputURLs(inputURLs, os.Stdout, os.Stderr, cfg.Verbose)
+	hasErr := safelinks.ProcessInputURLs(inputURLs, os.Stdout, os.Stderr, cfg.Verbose)
 
 	// Ensure unsuccessful error code if one encountered.
 	if hasErr {
