+++ internal/safelinks/doc.go
@@ -1,0 +1,3 @@
+// Package safelinks provides functionality for decoding Safe Links URLs to
+// their original values.
+package safelinks
