+++ internal/safelinks/output.go
@@ -5,7 +5,7 @@
 // Licensed under the MIT License. See LICENSE file in the project root for
 // full license information.
 
-package main
+package safelinks
 
 import (
 	"fmt"
